+++ src/main/kotlin/com/github/jensim/megamanipulator/actions/vcs/PrRouter.kt
@@ -1,5 +1,6 @@
 package com.github.jensim.megamanipulator.actions.vcs
 
+import com.github.jensim.megamanipulator.actions.NotificationsOperator
 import com.github.jensim.megamanipulator.actions.search.SearchResult
 import com.github.jensim.megamanipulator.actions.vcs.bitbucketserver.BitbucketServerClient
 import com.github.jensim.megamanipulator.actions.vcs.githubcom.GithubComClient
@@ -7,11 +8,14 @@ import com.github.jensim.megamanipulator.settings.CodeHostSettings
 import com.github.jensim.megamanipulator.settings.CodeHostSettings.BitBucketSettings
 import com.github.jensim.megamanipulator.settings.CodeHostSettings.GitHubSettings
 import com.github.jensim.megamanipulator.settings.SettingsFileOperator
+import com.intellij.notification.NotificationType.WARNING
+import java.util.concurrent.atomic.AtomicLong
 
 class PrRouter(
     private val settingsFileOperator: SettingsFileOperator,
     private val bitbucketServerClient: BitbucketServerClient,
     private val githubComClient: GithubComClient,
+    private val notificationsOperator: NotificationsOperator,
 ) {
 
     companion object {
@@ -21,26 +25,46 @@ class PrRouter(
                 settingsFileOperator = SettingsFileOperator.instance,
                 bitbucketServerClient = BitbucketServerClient.instance,
                 githubComClient = GithubComClient.instance,
+                notificationsOperator = NotificationsOperator.instance
             )
         }
     }
 
-    private fun resolve(searchHost: String, codeHost: String): CodeHostSettings? = settingsFileOperator.readSettings()
-        ?.searchHostSettings?.get(searchHost)?.codeHostSettings?.get(codeHost)
+    private val lastSettingsWarning = AtomicLong()
 
-    suspend fun addDefaultReviewers(pullRequest: PullRequestWrapper): PullRequestWrapper {
+    private fun resolve(searchHost: String, codeHost: String): CodeHostSettings? {
+        val resolved = settingsFileOperator.readSettings()
+            ?.searchHostSettings?.get(searchHost)?.codeHostSettings?.get(codeHost)
+        if (resolved == null) {
+            val last = lastSettingsWarning.get()
+            val current = System.currentTimeMillis()
+            if (last < (current - 100)) {
+                lastSettingsWarning.set(current)
+                notificationsOperator.show(
+                    title = "Missing config",
+                    body = "Failed finding config for '$searchHost'/'$codeHost'",
+                    type = WARNING
+                )
+            }
+        }
+        return resolved
+    }
+
+    suspend fun addDefaultReviewers(pullRequest: PullRequestWrapper): PullRequestWrapper? {
         val settings = resolve(pullRequest.searchHostName(), pullRequest.codeHostName())
         return when {
             settings is BitBucketSettings && pullRequest is BitBucketPullRequestWrapper -> bitbucketServerClient.addDefaultReviewers(settings, pullRequest)
             settings is GitHubSettings && pullRequest is GithubComPullRequestWrapper -> githubComClient.addDefaultReviewers(settings, pullRequest)
+            settings == null -> null
             else -> throw IllegalArgumentException("Unable to match config correctly")
         }
     }
 
-    suspend fun createPr(title: String, description: String, repo: SearchResult): PullRequestWrapper {
+    suspend fun createPr(title: String, description: String, repo: SearchResult): PullRequestWrapper? {
         return when (val settings = resolve(repo.searchHostName, repo.codeHostName)) {
             is BitBucketSettings -> bitbucketServerClient.createPr(title, description, settings, repo)
             is GitHubSettings -> githubComClient.createPr(title, description, settings, repo)
+            null -> null
             else -> throw IllegalArgumentException("Unable to match config correctly")
         }
     }
@@ -49,26 +73,28 @@ class PrRouter(
         return when (val settings = resolve(repo.searchHostName, repo.codeHostName)) {
             is BitBucketSettings -> bitbucketServerClient.createFork(settings, repo)
             is GitHubSettings -> githubComClient.createFork(settings, repo)
-            null -> throw IllegalArgumentException("Unable to match config correctly")
-        }
-    }
-
-    suspend fun updatePr(newTitle: String, newDescription: String, pullRequest: PullRequestWrapper): PullRequestWrapper {
-        val settings = resolve(pullRequest.searchHostName(), pullRequest.codeHostName())
-        return when {
-            settings is BitBucketSettings && pullRequest is BitBucketPullRequestWrapper -> bitbucketServerClient.updatePr(newTitle, newDescription, settings, pullRequest)
-            settings is GitHubSettings && pullRequest is GithubComPullRequestWrapper -> githubComClient.updatePr(newTitle, newDescription, settings, pullRequest)
+            null -> null
             else -> throw IllegalArgumentException("Unable to match config correctly")
         }
     }
 
-    suspend fun getAllPrs(searchHost: String, codeHost: String): List<PullRequestWrapper> {
-        return settingsFileOperator.readSettings()?.searchHostSettings?.get(searchHost)?.codeHostSettings?.get(codeHost)?.let {
+    suspend fun updatePr(newTitle: String, newDescription: String, pullRequest: PullRequestWrapper): PullRequestWrapper? {
+        val settings = resolve(pullRequest.searchHostName(), pullRequest.codeHostName())
+        return when {
+            settings is BitBucketSettings && pullRequest is BitBucketPullRequestWrapper -> bitbucketServerClient.updatePr(newTitle, newDescription, settings, pullRequest)
+            settings is GitHubSettings && pullRequest is GithubComPullRequestWrapper -> githubComClient.updatePr(newTitle, newDescription, settings, pullRequest)
+            settings == null -> null
+            else -> throw IllegalArgumentException("Unable to match config correctly")
+        }
+    }
+
+    suspend fun getAllPrs(searchHost: String, codeHost: String): List<PullRequestWrapper>? {
+        return resolve(searchHost, codeHost)?.let {
             when (it) {
                 is BitBucketSettings -> bitbucketServerClient.getAllPrs(searchHost, codeHost, it)
                 is GitHubSettings -> githubComClient.getAllPrs(searchHost, codeHost, it)
             }
-        } ?: throw IllegalArgumentException("No config!")
+        }
     }
 
     suspend fun closePr(dropForkOrBranch: Boolean, pullRequest: PullRequestWrapper) {
@@ -76,35 +102,37 @@ class PrRouter(
         when {
             settings is BitBucketSettings && pullRequest is BitBucketPullRequestWrapper -> bitbucketServerClient.closePr(dropForkOrBranch, settings, pullRequest)
             settings is GitHubSettings && pullRequest is GithubComPullRequestWrapper -> githubComClient.closePr(dropForkOrBranch, settings, pullRequest)
+            settings == null -> Unit
             else -> throw IllegalArgumentException("Unable to match config correctly")
         }
     }
 
-    suspend fun getPrivateForkReposWithoutPRs(searchHost: String, codeHost: String): List<RepoWrapper> {
-        return settingsFileOperator.readSettings()?.searchHostSettings?.get(searchHost)?.codeHostSettings?.get(codeHost)?.let {
+    suspend fun getPrivateForkReposWithoutPRs(searchHost: String, codeHost: String): List<RepoWrapper>? {
+        return resolve(searchHost, codeHost)?.let {
             when (it) {
                 is BitBucketSettings -> bitbucketServerClient.getPrivateForkReposWithoutPRs(searchHost, codeHost, it)
                 is GitHubSettings -> githubComClient.getPrivateForkReposWithoutPRs(searchHost, codeHost, it)
             }
-        } ?: throw IllegalArgumentException("No config!")
+        }
     }
 
     suspend fun deletePrivateRepo(fork: RepoWrapper) {
-        settingsFileOperator.readSettings()?.searchHostSettings?.get(fork.getSearchHost())?.codeHostSettings?.get(fork.getCodeHost())?.let { it ->
+        resolve(fork.getSearchHost(), fork.getCodeHost())?.let { settings ->
             when {
-                it is BitBucketSettings && fork is BitBucketRepoWrapping -> bitbucketServerClient.deletePrivateRepo(fork, it)
-                it is GitHubSettings && fork is GithubComRepoWrapping -> githubComClient.deletePrivateRepo(fork, it)
+                settings is BitBucketSettings && fork is BitBucketRepoWrapping -> bitbucketServerClient.deletePrivateRepo(fork, settings)
+                settings is GitHubSettings && fork is GithubComRepoWrapping -> githubComClient.deletePrivateRepo(fork, settings)
                 else -> throw IllegalArgumentException("Unable to match config correctly")
             }
-        } ?: throw IllegalArgumentException("No config!")
+        }
     }
 
-    suspend fun getRepo(searchResult: SearchResult): RepoWrapper {
-        return settingsFileOperator.readSettings()?.searchHostSettings?.get(searchResult.searchHostName)?.codeHostSettings?.get(searchResult.codeHostName)?.let {
+    suspend fun getRepo(searchResult: SearchResult): RepoWrapper? {
+        val settings = resolve(searchResult.searchHostName, searchResult.codeHostName)
+        return settings?.let {
             when (it) {
                 is BitBucketSettings -> bitbucketServerClient.getRepo(searchResult, it)
                 is GitHubSettings -> githubComClient.getRepo(searchResult, it)
             }
-        } ?: throw IllegalArgumentException("No config!")
+        }
     }
 }